+++ Medidata.Cloud.Tsdv.Loader/CellTypeConverters/BooleanConverter.cs
@@ -1,6 +1,6 @@
 using DocumentFormat.OpenXml.Spreadsheet;
 
-namespace Medidata.Cloud.Tsdv.Loader.CellValueConverters
+namespace Medidata.Cloud.Tsdv.Loader.CellTypeConverters
 {
     internal class BooleanConverter : CellValueBaseConverter<bool>
     {
