+++ Medidata.Cloud.Tsdv.Loader/CellTypeConverters/CellTypeConverterManager.cs
@@ -2,7 +2,7 @@ using System;
 using System.Linq;
 using DocumentFormat.OpenXml.Spreadsheet;
 
-namespace Medidata.Cloud.Tsdv.Loader.CellValueConverters
+namespace Medidata.Cloud.Tsdv.Loader.CellTypeConverters
 {
     internal class CellTypeConverterManager
     {
