+++ Medidata.Cloud.Tsdv.Loader/CellTypeConverters/ICellValueConverter.cs
@@ -1,7 +1,7 @@
 using System;
 using DocumentFormat.OpenXml.Spreadsheet;
 
-namespace Medidata.Cloud.Tsdv.Loader.CellValueConverters
+namespace Medidata.Cloud.Tsdv.Loader.CellTypeConverters
 {
     public interface ICellValueConverter
     {
