+++ Medidata.Cloud.Tsdv.Loader/CellTypeConverters/InlineStringConverter.cs
@@ -1,17 +1,17 @@
 using DocumentFormat.OpenXml.Spreadsheet;
 
-namespace Medidata.Cloud.Tsdv.Loader.CellValueConverters
+namespace Medidata.Cloud.Tsdv.Loader.CellTypeConverters
 {
-    internal class StringConverter : CellValueBaseConverter<string>
+    internal class InlineStringConverter : CellValueBaseConverter<string>
     {
-        public StringConverter() : base(CellValues.String)
+        public InlineStringConverter()
+            : base(CellValues.InlineString)
         {
         }
 
         protected override string GetCellValueImpl(string csharpValue)
         {
             return csharpValue;
-            ;
         }
 
         protected override string GetCSharpValueImpl(string cellValue)
