+++ Medidata.Cloud.Tsdv.Loader/CellTypeConverters/NullableBooleanConverter.cs
@@ -1,6 +1,6 @@
 using DocumentFormat.OpenXml.Spreadsheet;
 
-namespace Medidata.Cloud.Tsdv.Loader.CellValueConverters
+namespace Medidata.Cloud.Tsdv.Loader.CellTypeConverters
 {
     internal class NullableBooleanConverter : CellValueBaseConverter<bool?>
     {
