+++ Medidata.Cloud.Tsdv.Loader/CellTypeConverters/NumberConverter.cs
@@ -1,6 +1,6 @@
 using DocumentFormat.OpenXml.Spreadsheet;
 
-namespace Medidata.Cloud.Tsdv.Loader.CellValueConverters
+namespace Medidata.Cloud.Tsdv.Loader.CellTypeConverters
 {
     internal abstract class NumberConverter<T> : CellValueBaseConverter<T>
     {
