+++ Medidata.Cloud.Tsdv.Loader/CellTypeConverters/StringConverter.cs
@@ -1,17 +1,17 @@
 using DocumentFormat.OpenXml.Spreadsheet;
 
-namespace Medidata.Cloud.Tsdv.Loader.CellValueConverters
+namespace Medidata.Cloud.Tsdv.Loader.CellTypeConverters
 {
-    internal class InlineStringConverter : CellValueBaseConverter<string>
+    internal class StringConverter : CellValueBaseConverter<string>
     {
-        public InlineStringConverter()
-            : base(CellValues.InlineString)
+        public StringConverter() : base(CellValues.String)
         {
         }
 
         protected override string GetCellValueImpl(string csharpValue)
         {
             return csharpValue;
+            ;
         }
 
         protected override string GetCSharpValueImpl(string cellValue)
